+++ src/utils/k-permutation.h
@@ -3,6 +3,6 @@
 
 #include <stddef.h> /* For size_t */
 
-unsigned int AInext_k_permutation(unsigned int *ar, size_t n, int k);
+unsigned int AInext_k_permutation(unsigned int *ar, size_t n, size_t k);
 
 #endif
